requirements.txt
CMake
opencv-python
dlib
face_recognition
pillow